enhancement (playground/search-pg): move search-results component in search-pg, update paths and clean up
