+++ src/app/knora-ui-examples/search-demo/search-panel/search-panel.component.html
@@ -17,10 +17,3 @@
         <img src="data/documentation/search/search-panel-example.png" class="center" />
     </div>
 </mat-card>
-
-<!-- <h3>Demo/Test - search results</h3>
-<mat-card class="example-viewer">
-    <kui-search [route]="'/modules/search/search'"></kui-search>
-</mat-card>
-
-<router-outlet></router-outlet> -->
