+++ src/app/playground/search-pg/search-pg.component.html
@@ -1,15 +1,24 @@
-<div class="search-content">
+<div class="search-bars">
 
   <h3>Old search</h3>
-  <kui-search [route]="'/modules/search/search'"></kui-search>
+  <kui-search [route]="'/playground/search'"></kui-search>
   <br>
   <h3>Fulltext search</h3>
-  <kui-fulltext-search [route]="'/modules/search/search'"></kui-fulltext-search>
+  <kui-fulltext-search [route]="'/playground/search'"></kui-fulltext-search>
   <br>
   <h3>Extended search</h3>
-  <kui-extended-search [route]="'/modules/search/search'"></kui-extended-search>
+  <kui-extended-search [route]="'/playground/search'"></kui-extended-search>
   <br>
   <h3>Search panel</h3>
-  <kui-search-panel [route]="'/modules/search/search'"></kui-search-panel>
+  <kui-search-panel [route]="'/playground/search'"></kui-search-panel>
+
+</div>
+
+<mat-divider></mat-divider>
+
+<div class="search-results">
+
+  <h3>Search results</h3>
+  <router-outlet></router-outlet>
 
 </div>
