+++ src/app/playground/search-pg/search-pg.component.scss
@@ -1,3 +1,6 @@
-.search-content {
+.search-bars {
+    margin: 30px;
+}
+.search-results {
     margin: 30px;
 }
