+++ src/app/playground/search-pg/search-result/search-result.component.html
